updating all components

+++ apps/controllers/jadwal_controller.py
@@ -1,29 +1,41 @@
+from fastapi import HTTPException, Depends
 from sqlalchemy.orm import Session
-from apps.models.jadwal import Jadwal
+from apps.database import get_db
+from apps.models.jadwal import Jadwal 
 
-def create_jadwal(db: Session, jadwal_data: Jadwal):
+def create_jadwal(jadwal_data: Jadwal, db: Session = Depends(get_db)):
     db_jadwal = Jadwal(**jadwal_data)
     db.add(db_jadwal)
     db.commit()
     db.refresh(db_jadwal)
     return db_jadwal
 
-def get_jadwal(db: Session, kode_jadwal: str):
-    return db.query(Jadwal).filter(Jadwal.kode_jadwal == kode_jadwal).first()
+def get_jadwal(kd_jadwal: str, db: Session = Depends(get_db)):
+    jadwal = db.query(Jadwal).filter(Jadwal.kd_jadwal == kd_jadwal).first()
+    if jadwal is None:
+        raise HTTPException(status_code=404, detail="Jadwal not found")
+    return jadwal
 
-def get_jadwals(db: Session):
-    return db.query(Jadwal).all()
+def get_all_jadwal(skip: int = 0, limit: int = 10, db: Session = Depends(get_db)):
+    jadwals = db.query(Jadwal).offset(skip).limit(limit).all()
+    return jadwals
 
-def update_jadwal(db: Session, kode_jadwal: str, jadwal_data: Jadwal):
-    db_jadwal = db.query(Jadwal).filter(Jadwal.kode_jadwal == kode_jadwal).first()
-    for key, value in jadwal_data.items():
+def update_jadwal(jadwal_data: Jadwal, kd_jadwal: str, db: Session = Depends(get_db)):
+    db_jadwal = db.query(Jadwal).filter(Jadwal.kd_jadwal == kd_jadwal).first()
+    if db_jadwal is None:
+        raise HTTPException(status_code=404, detail="Jadwal not found")
+    
+    for key, value in jadwal_data.dict().items():
         setattr(db_jadwal, key, value)
     db.commit()
     db.refresh(db_jadwal)
     return db_jadwal
 
-def delete_jadwal(db: Session, kode_jadwal: str):
-    db_jadwal = db.query(Jadwal).filter(Jadwal.kode_jadwal == kode_jadwal).first()
+def delete_jadwal(kd_jadwal: str, db: Session = Depends(get_db)):
+    db_jadwal = db.query(Jadwal).filter(Jadwal.kd_jadwal == kd_jadwal).first()
+    if db_jadwal is None:
+        raise HTTPException(status_code=404, detail="Jadwal not found")
+
     db.delete(db_jadwal)
     db.commit()
     return {"message": "Jadwal deleted successfully"}
